requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
SQLAlchemy
psycopg2-binary
line-bot-sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
APScheduler
cachetools
python-dateutil
python-dotenv
pytz
qrcode
Pillow
ImageHash
num2words
pandas
openpyxl
blinker
click
certifi
charset-normalizer
colorama
exceptiongroup
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
idna
importlib-metadata
iniconfig
isodate
liffpy
more-itertools
numpy
oauthlib
packaging
pluggy
protobuf
PyJWT
pyparsing
pytest
PyYAML
requests-oauthlib
rfc3986
six
sniffio
tblib
tomli
tzdata
urllib3
zipp
Flask-Migrate
sentry-sdk
geopy
flask_cors
weasyprint